refactor: consolidate all models into one module
The models were getting messy and cross-contamainted, so rather just
place them in one module
