+++ github-api/src/lib.rs
@@ -1,12 +1,13 @@
 pub mod api;
 pub mod error;
+pub mod github_event;
 mod github_provider;
 pub mod graphql;
 mod macros;
 pub mod models;
 pub mod models_plus;
 pub mod provider_traits;
-pub mod webhook_events;
 pub mod wrappers;
 
+pub use github_event::GithubEvent;
 pub use github_provider::GithubProvider;
