+++ github-api/src/models/mod.rs
@@ -1,39 +1,59 @@
 mod app;
 mod branch;
+mod comment_events;
 mod comments;
 mod commits;
 mod common;
 mod date_time;
+mod events_common;
 mod git;
 mod installation;
 mod integration;
+mod issue_event;
 mod issues;
+mod label_event;
 mod labels;
 mod links;
 mod milestone;
 mod organization;
+mod ping_event;
 mod pull_request;
+mod pull_request_event;
 mod pull_request_review;
 mod pull_request_review_comment;
+mod pull_request_review_event;
+mod push_event;
 mod repository;
+mod status_check_events;
+mod status_event;
 mod team;
 mod user;
 
 pub use app::*;
 pub use branch::Branch;
+pub use comment_events::*;
 pub use comments::*;
 pub use commits::*;
 pub use common::*;
 pub use date_time::DateTime;
+pub use events_common::*;
 pub use git::GitReference;
 pub use installation::*;
+pub use issue_event::*;
 pub use issues::Issue;
-pub use labels::Label;
+pub use label_event::*;
+pub use labels::{Label, *};
 pub use links::Links;
 pub use organization::Organization;
+pub use ping_event::*;
 pub use pull_request::*;
+pub use pull_request_event::*;
 pub use pull_request_review::*;
 pub use pull_request_review_comment::*;
+pub use pull_request_review_event::*;
+pub use push_event::*;
 pub use repository::*;
+pub use status_check_events::*;
+pub use status_event::*;
 pub use team::SimpleTeam;
 pub use user::*;
