--- github-api/src/webhook_events/mod.rs
@@ -1,9 +1,0 @@
-mod github_event;
-mod issue_event;
-mod models;
-mod pr_event;
-mod status_check_events;
-
-pub use github_event::GithubEvent;
-pub use models::*;
-pub use status_check_events::*;
