+++ server/src/pub_sub/actor.rs
@@ -1,7 +1,7 @@
 use std::sync::{Arc, RwLock};
 
 use actix::{Actor, Context, Handler, Running, SystemService};
-use github_pilot_api::webhook_events::GithubEvent;
+use github_pilot_api::GithubEvent;
 use log::*;
 
 use crate::{
